requirements.txt
django==5.2.6
pyyaml==6.0.2